[fix] 주차 삭제 시 해당 주차와 관련된 출석 정보가 삭제되지 않는 오류를 수정한다 (#216)
* fix: 특정 주차 삭제 시 해당 주차와 관련된 출석 정보 삭제

+++ src/main/java/com/kgu/studywithme/study/infra/query/StudySimpleQueryRepositoryImpl.java
@@ -137,11 +137,7 @@ public class StudySimpleQueryRepositoryImpl implements StudySimpleQueryRepositor
                 .orderBy(weekly.week.desc())
                 .fetch();
 
-        if (weeks.size() == 0) {
-            return 1;
-        }
-
-        return weeks.get(0) + 1;
+        return weeks.size() == 0 ? 1 : weeks.get(0) + 1;
     }
 
     @Override
@@ -176,17 +172,26 @@ public class StudySimpleQueryRepositoryImpl implements StudySimpleQueryRepositor
                 )
                 .fetchOne();
 
-        query.delete(submit)
-                .where(submit.week.id.eq(weekId))
-                .execute();
+        if (weekId != null) {
+            query.delete(submit)
+                    .where(submit.week.id.eq(weekId))
+                    .execute();
 
-        query.delete(attachment)
-                .where(attachment.week.id.eq(weekId))
-                .execute();
+            query.delete(attachment)
+                    .where(attachment.week.id.eq(weekId))
+                    .execute();
 
-        query.delete(weekly)
-                .where(weekly.id.eq(weekId))
-                .execute();
+            query.delete(attendance)
+                    .where(
+                            attendance.study.id.eq(studyId),
+                            attendance.week.eq(week)
+                    )
+                    .execute();
+
+            query.delete(weekly)
+                    .where(weekly.id.eq(weekId))
+                    .execute();
+        }
     }
 
     private BooleanExpression participateStatusEq(ParticipantStatus status) {
